+++ source/modulo_components/src/LifecycleComponent.cpp
@@ -9,7 +9,7 @@ namespace modulo_components {
 
 LifecycleComponent::LifecycleComponent(const rclcpp::NodeOptions& node_options, const std::string& fallback_name)
     : LifecycleNode(
-        utilities::parse_node_name(node_options, fallback_name), utilities::modify_parameter_overrides(node_options)),
+        modulo_utils::parse_node_name(node_options, fallback_name), utilities::modify_parameter_overrides(node_options)),
       ComponentInterface(std::make_shared<rclcpp::node_interfaces::NodeInterfaces<ALL_RCLCPP_NODE_INTERFACES>>(
           LifecycleNode::get_node_base_interface(), LifecycleNode::get_node_clock_interface(),
           LifecycleNode::get_node_graph_interface(), LifecycleNode::get_node_logging_interface(),