+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Release Versions:
 - feat(components): only publish predicates on value change (#123)
 - feat(controllers): only publish predicates on value change (#124)
 - fix(controllers): keep parameter values in sync with ros (#125)
+- feat: use v2.0.0-rc1-jazzy image (#126)
 
 ## 4.2.2